requirements/requirements.txt
customtkinter
requests
instaloader
pillow